requirements.txt
streamlit==1.33.0
requests==2.31.0
tensorflow==2.15.0
keras==2.15.0
pandas==2.2.2
plotly==5.20.0
scikit-learn==1.4.2
numpy==1.26.4
newspaper3k==0.2.8
python-dotenv==1.0.0
lxml_html_clean==0.4.2
